cron/requirements.txt
asgiref==3.4.1
astroid==2.6.2
awscli==1.22.7
boto3==1.20.7
botocore==1.23.7
colorama==0.4.3
cycler==0.11.0
Django==3.2.5
django-crontab==0.7.1
django-storages==1.12.3
djangorestframework==3.12.4
docutils==0.15.2
fonttools==4.28.1
gunicorn==20.1.0
isort==5.9.1
jmespath==0.10.0
kiwisolver==1.3.2
lazy-object-proxy==1.6.0
matplotlib==3.5.0
mccabe==0.6.1
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
pyasn1==0.4.8
pylint==2.9.3
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.2
urllib3==1.26.7
wrapt==1.12.1
